refactor: update type hints for better clarity and consistency in playlist and video progress handling

+++ src/Domain/Playlists/Exceptions/DiskUsageException.php
@@ -9,7 +9,7 @@ use Illuminate\Support\Number;
 
 class DiskUsageException extends Exception
 {
-    public static function exceededLimit(string $path, float|int $limit = 0): self
+    public static function exceededLimit(string $path, int|float $limit = 0): self
     {
         $freeSpace = Number::fileSize($limit);
